+++ normalized-cache-incubating/src/jsCommonMain/kotlin/com/apollographql/cache/normalized/internal/Lock.jsCommon.kt
@@ -1,0 +1,11 @@
+package com.apollographql.cache.normalized.internal
+
+internal actual fun Lock(): Lock = object : Lock {
+  override fun <T> read(block: () -> T): T {
+    return block()
+  }
+
+  override fun <T> write(block: () -> T): T {
+    return block()
+  }
+}